Make testing easier
